Add files via upload
The objective of these codes is to analyze the diversity and quality of applicants and based on that provide useful suggestions to reduce bias and increase the diversity. Essentially, we want to find out what language affects diversity and quality and whether salary has any association with diversity and in turn affects the quality.